adds xweb share config validation points, time.ParseDuration
- split xweb/config.go into multiple files by type
- allow api factories to validate shared config
- use time.ParseDuration for duration config values in core xweb options
- add interface documentation for api factories
- merge http.request context values for xweb
