+++ controller/xweb/registry.go
@@ -78,6 +78,14 @@ type WebHandler interface {
 // instances based on need. This interface allows WebHandler logic to be reused across multiple xweb.Server's while
 // delegating the instance management to the factory.
 type WebHandlerFactory interface {
+	// Binding returns the string binding value used in configurations
 	Binding() string
+
+	// New creates a new instances of this factories API
 	New(webListener *WebListener, options map[interface{}]interface{}) (WebHandler, error)
+
+	// Validate is used for factory level configuration checks. It is not meant for individual instance validation.
+	// It is meant for validating configuration that all instances generated by a factory share (i.e. sanity across
+	// instances or configuration from another section).
+	Validate(config *Config) error
 }
